+++ packages/devon/src/commands/index.ts
@@ -1,0 +1,6 @@
+export * as down from './down';
+export * as exec from './exec';
+export * as init from './init';
+export * as install from './install';
+export * as run from './run';
+export * as switch from './switch';