+++ packages/devon/src/helpers/services/types.ts
@@ -1,0 +1,9 @@
+import type { Service, DefinitionFile, ServiceConfig } from '@/types';
+
+export interface Result {
+    definition: DefinitionFile;
+    services: Record<string, {
+        definition: Service;
+        config: ServiceConfig;
+    }>;
+}