+++ src/commands/index.ts
@@ -1,6 +1,0 @@
-export * from './down';
-export * from './exec';
-export * from './init';
-export * from './install';
-export * from './run';
-export * from './switch';